+++ services/pull/protection_check.go
@@ -2,18 +2,17 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package branchprotection
+package pull
 
 import (
 	"context"
 
 	"code.gitea.io/gitea/models"
 	user_model "code.gitea.io/gitea/models/user"
-	pull_service "code.gitea.io/gitea/services/pull"
 )
 
-// Check if pull is mergable
-func Check(ctx context.Context, doer *user_model.User, perm *models.Permission, pr *models.PullRequest, manuallMerge, force bool) error {
+// CheckPullProtection check if the pull mergable based on all conditions (branch protection, merge options, ...)
+func CheckPullProtection(ctx context.Context, doer *user_model.User, perm *models.Permission, pr *models.PullRequest, manuallMerge, force bool) error {
 	if pr.HasMerged {
 		return ErrHasMerged{}
 	}
@@ -24,7 +23,7 @@ func Check(ctx context.Context, doer *user_model.User, perm *models.Permission, 
 		return ErrIsClosed{}
 	}
 
-	if allowedMerge, err := pull_service.IsUserAllowedToMerge(pr, *perm, doer); err != nil {
+	if allowedMerge, err := IsUserAllowedToMerge(pr, *perm, doer); err != nil {
 		return err
 	} else if !allowedMerge {
 		return ErrUserNotAllowedToMerge{}
@@ -43,7 +42,7 @@ func Check(ctx context.Context, doer *user_model.User, perm *models.Permission, 
 		return ErrNotMergableState{}
 	}
 
-	if err := pull_service.CheckPRReadyToMerge(pr, false); err != nil {
+	if err := CheckPRReadyToMerge(pr, false); err != nil {
 		if models.IsErrNotAllowedToMerge(err) {
 			if force {
 				if isRepoAdmin, err := models.IsUserRepoAdmin(pr.BaseRepo, doer); err != nil {
@@ -57,7 +56,7 @@ func Check(ctx context.Context, doer *user_model.User, perm *models.Permission, 
 		}
 	}
 
-	if _, err := pull_service.IsSignedIfRequired(pr, doer); err != nil {
+	if _, err := IsSignedIfRequired(pr, doer); err != nil {
 		return err
 	}
